Fix #1 Add small test set
This removes test data dependencies of sufam on files on saba2. It uses a
subset now of the original bams and only chr17 of the human reference genome is
included in the repo. This makes the repo larger, but since sufam is packaged
on pip only developers will notice the difference and they should run the tests
anyway.